feat: a json parsing test
