+++ src/TMTProductizer.UnitTests/MockUtils.cs
@@ -1,0 +1,28 @@
+
+namespace TMTProductizer.UnitTests.Mocks;
+
+public static class MockUtils
+{
+
+    public static string GetTMTTestResponse()
+    {
+        try
+        {
+            // Open the text file using a stream reader.
+            using (var sr = new StreamReader("../../src/TMTProductizer.UnitTests/Mocks/testTMTResponse.json"))
+            {
+                // Read the stream to a string, and write the string to the console.
+                var line = sr.ReadToEnd();
+                return line;
+            }
+        }
+        catch (IOException e)
+        {
+            Console.WriteLine("The file could not be read:");
+            Console.WriteLine(e.Message);
+            throw e;
+
+        }
+
+    }
+}
