+++ src/TMTProductizer.UnitTests/Mocks/TestModel.cs
@@ -1,0 +1,6 @@
+namespace TMTProductizer.UnitTests.Mocks;
+
+public class TestModel
+{
+    public string Name { get; set; } = null!;
+}
