+++ src/TMTProductizer.UnitTests/Services/JobServiceTests.cs
@@ -6,6 +6,7 @@ using Moq.Protected;
 using TMTProductizer.Models;
 using TMTProductizer.Services;
 using TMTProductizer.Services.AWS;
+using TMTProductizer.UnitTests.Mocks;
 using TMTProductizer.Utils;
 
 namespace TMTProductizer.UnitTests.Services;
@@ -15,7 +16,7 @@ public class JobServiceTests
     [Test]
     public void TryingToFindJob_WithTmtApiUp_ReturnsListWithData()
     {
-        string TmtJson = GetTMTTestResponse();
+        string TmtJson = MockUtils.GetTMTTestResponse();
         var handler = new Mock<HttpMessageHandler>(MockBehavior.Strict);
         handler.Protected()
             .Setup<Task<HttpResponseMessage>>(
@@ -48,7 +49,7 @@ public class JobServiceTests
             Paging = new PagingOptions
             {
                 Limit = 1,
-                Offset = 20
+                Offset = 0
             }
         };
 
@@ -106,26 +107,5 @@ public class JobServiceTests
         result.Should().NotBeNull();
         result.Result.Should().BeOfType<(List<Job>, long)>();
         result.Result.jobs.Count.Should().Be(0);
-    }
-
-
-    private string GetTMTTestResponse()
-    {
-        try
-        {
-            // Open the text file using a stream reader.
-            using (var sr = new StreamReader("../../src/TMTProductizer.UnitTests/Mocks/testTMTResponse.json"))
-            {
-                // Read the stream to a string, and write the string to the console.
-                var line = sr.ReadToEnd();
-                return line;
-            }
-        }
-        catch (IOException e)
-        {
-            Console.WriteLine("The file could not be read:");
-            Console.WriteLine(e.Message);
-            throw e;
-        }
     }
 }
